+++ src/main/java/edu/unicauca/apliweb/persistence/jpa/CompraJpaController.java
@@ -109,7 +109,7 @@ public class CompraJpaController implements Serializable {
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                Long id = compra.getIdCompra();
+                Integer id = compra.getIdCompra();
                 if (findCompra(id) == null) {
                     throw new NonexistentEntityException("The compra with id " + id + " no longer exists.");
                 }
@@ -122,7 +122,7 @@ public class CompraJpaController implements Serializable {
         }
     }
 
-    public void destroy(Long id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -177,7 +177,7 @@ public class CompraJpaController implements Serializable {
         }
     }
 
-    public Compra findCompra(Long id) {
+    public Compra findCompra(Integer id) {
         EntityManager em = getEntityManager();
         try {
             return em.find(Compra.class, id);
