+++ src/main/java/edu/unicauca/apliweb/persistence/jpa/ProveedorJpaController.java
@@ -97,7 +97,7 @@ public class ProveedorJpaController implements Serializable {
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                Long id = proveedor.getCodProveedor();
+                Integer id = proveedor.getCodProveedor();
                 if (findProveedor(id) == null) {
                     throw new NonexistentEntityException("The proveedor with id " + id + " no longer exists.");
                 }
@@ -110,7 +110,7 @@ public class ProveedorJpaController implements Serializable {
         }
     }
 
-    public void destroy(Long id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -160,7 +160,7 @@ public class ProveedorJpaController implements Serializable {
         }
     }
 
-    public Proveedor findProveedor(Long id) {
+    public Proveedor findProveedor(Integer id) {
         EntityManager em = getEntityManager();
         try {
             return em.find(Proveedor.class, id);
